myBatis/src/main/java/lesson1_ibatis/service/StudentService.java
package lesson1_ibatis.service;/**
 * @author Sergey Zhernovoy
 * create on 16.11.2017.
 */

import lesson1_ibatis.domain.Student;
import lesson1_ibatis.mappers.StudentMapper;
import lesson1_ibatis.util.MyBatisSqlSessionFactory;
import org.apache.ibatis.session.SqlSession;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

import java.util.List;

public class StudentService implements StudentMapper {
    private static final Logger logger = LoggerFactory.getLogger(StudentService.class);

    @Override
    public List<Student> findAllStudents() {
        SqlSession sqlSession = MyBatisSqlSessionFactory.openSession();
        try {
            StudentMapper studentMapper = sqlSession.getMapper(StudentMapper.class);
            return studentMapper.findAllStudents();
        } finally {
            sqlSession.close();
        }
    }

    @Override
    public Student findStudentById(Integer id) {
        logger.debug("Select stident by ID:{}", id);
        SqlSession sqlSession = MyBatisSqlSessionFactory.openSession();
        try {
            StudentMapper studentMapper = sqlSession.getMapper(StudentMapper.class);
            return studentMapper.findStudentById(id);
        } finally {
            sqlSession.close();
        }
    }

    @Override
    public void insertStudent(Student student) {
        SqlSession sqlSession = MyBatisSqlSessionFactory.openSession();
        try {
            StudentMapper studentMapper = sqlSession.getMapper(StudentMapper.class);
            studentMapper.insertStudent(student);
            sqlSession.commit();
        } finally {
            sqlSession.close();
        }
    }
}

    